+++ .eslintrc.js
@@ -1,3 +1,6 @@
 module.exports = {
   extends: ["plugin:prettier/recommended", "plugin:node/recommended"],
+  parserOptions: {
+    ecmaVersion: 2020,
+  },
 };